Calculations Complete
Calculations are complete now. But i have to make some changes for the
google timezone API. Istead of giving +2 hours sometime reports
+1:56:56 and gives +1 rawOffset.

+++ Gokturk/Observer.h
@@ -18,6 +18,6 @@
 @property NSString *timeZoneId;
 @property NSString *timeZoneName;
 
--(void) createObserverWithLatitude:(double) newLatitude andLongitude:(double) newLongitude andHeading:(int) newHeading andTimeZoneId:(NSString *) newTimeZoneId andTimeZoneName:(NSString *) newTimeZoneName andDSTOffset:(double) newDSTOffset andRawOffset:(double)newRawOffset ;
+-(void) createObserverWithLatitude:(double) newLatitude andLongitude:(double) newLongitude andHeading:(int) newHeading andTimeZoneId:(NSString *) newTimeZoneId andTimeZoneName:(NSString *) newTimeZoneName andDSTOffset:(int) newDSTOffset andRawOffset:(int)newRawOffset ;
 
 @end
